backend/pyproject.toml
[project]
name = "clawhub-backend"
version = "0.1.0"
description = "ClawHub - Container Registry Platform Backend"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "ClawHub Team", email = "contact@clawhub.io" }
]
keywords = ["docker", "registry", "container", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web Framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    
    # Cache
    "redis>=5.2.0",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    
    # Storage
    "minio>=7.2.0",
    "boto3>=1.35.0",
    
    # HTTP Client
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
    
    # Background Tasks
    "celery>=5.4.0",
    
    # Monitoring
    "prometheus-client>=0.21.0",
    "sentry-sdk[fastapi]>=2.18.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "factory-boy>=3.3.0",
    "faker>=30.8.0",
    
    # Linting & Formatting
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    
    # Type Stubs
    "types-python-jose>=3.3.0",
    "types-passlib>=1.7.0",
    "types-redis>=4.6.0",
]

[project.scripts]
clawhub = "app.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/clawhub"
Documentation = "https://docs.clawhub.io"
Repository = "https://github.com/your-org/clawhub"
Issues = "https://github.com/your-org/clawhub/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# -------------------------------------------------
# Ruff 配置
# -------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PT",  # flake8-pytest-style
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function call in argument defaults
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG001"]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# -------------------------------------------------
# Mypy 配置
# -------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# -------------------------------------------------
# Pytest 配置
# -------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
fail_under = 80
